Blacklisted PyFunceble as they are going closed source and only for money.
I'm working on the authentication + password/machine account mechanism but I hope to soon make this project a thing for some of the thing I work on ...

It won't be Open Source. And I may charge a few bucks if I choose to make that thing available to communities ...

What the thing can do:

- Collect subject test result
- Collect subject
- Collect subject from a submitted remote file
- Schedule Subject to test next and make them available.
- A custom client will run on a machine an communicate with the API to get the next subject to text. It will then test the thing (using the PyFunceble Python API) then submit the result to the platform.

It's like a distributed system but that is not perfect yet haha. The only thing to do over time, is just to add new machines/node, and wait for subject to popup through remote files, user search or whatever source I may add in there later hihi

Oh and the thing provide an api for each submitted file ... So it may become a thing for people who want their list tested, and available through an API ...